+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - add live demo link to README.md
+- center table headers
+
 
 ## [0.1.1] - 2025-10-21